Fitting function implemented in utilities.py > fitting function saves a figure > fitting function does not save the data > improved plotting routines > parameters visualizations needs improvement
