+++ Python/main.py
@@ -5,7 +5,6 @@ import torch
 
 # impor numeric libraries
 import numpy as np
-import matplotlib.pyplot as plt
 
 # import Custom Modules
 import parameters as pm
@@ -71,71 +70,14 @@ mesh = grid.get_points().to(device)
 # torch.requires_grad_() tells autograd to record operations on this tensor
 x_grid = mesh.clone().unsqueeze(1).requires_grad_().type(torch.complex128)
 
-#%% Fitting initial condition
-from tqdm import tqdm
-
-learning_rate = 1e-3
-optimizer = torch.optim.Adam(params=model.parameters(), lr=learning_rate)
-
-# Create x_target as a leaf tensor
-x_target = torch.tensor(mesh, dtype=torch.complex128, requires_grad=False).clone().unsqueeze(1)
-# Target wavefunction
-target = (1/np.pi)**(1/4) * np.exp(-0.5 * (x_target)**2)
-
-def loss_fn(psi, target):
-    # Complex Mean Squared Error (CMSE)
-    real_loss = torch.mean((psi.real - target.real)**2)
-    imag_loss = torch.mean((psi.imag - target.imag)**2)
-    return real_loss + imag_loss
-
-scheduler = torch.optim.lr_scheduler.StepLR(optimizer, step_size=1e3, gamma=0.5)  # Reduce LR by 50% every 100 steps
-
-# Training
-loss_lst = []
-epochs = 10000
-with tqdm(total=epochs, disable=not pm.progress_bar) as pbar:
-    for i in range(epochs):   
-        # wavefunction
-        psi = model(x_target)
-        # Compute the loss
-        loss = loss_fn(psi, target) 
-        # Perform backpropagation and optimization
-        optimizer.zero_grad()   # Initialize gradients to zero at each epoch
-        loss.backward(retain_graph=False)         # Compute gradients
-        optimizer.step()        # Update parameters
-        
-        # scheduler.step()  # Adjust learning rate
-
-        # Optionally: print the loss for debugging
-        pbar.set_postfix_str(f"Epoch {i}, Loss: {loss.item():.2e}")
-        pbar.update(1)  
-
-        loss_lst.append(loss.detach().numpy())
-
-#%% Visualization
-fig = plt.figure()  # Create the main figure
-
-# Primary Y-Axis: Target and Model
-ax = fig.add_subplot(111, facecolor="none")
-ax.plot(mesh, np.abs(target.detach().numpy())**2, label='Target') 
-ax.plot(mesh, np.abs(model(x_target).detach().numpy())**2, label='Model')
-ax.set_xlabel('Mesh (Primary X-Axis)')  # Primary x-axis label
-ax.set_ylabel('Amplitude Squared')  # Primary y-axis label
-ax.legend(loc='upper right')
-
-# Add Loss plot to the primary axis (optional)
-ax2 = fig.add_subplot(111, facecolor="none")  # Secondary y-axis for loss
-ax2.plot(range(epochs), loss_lst, label='Loss', color='red', linestyle='dashed')
-ax2.set_yscale('log')
-ax2.xaxis.tick_top()
-ax2.xaxis.set_label_position('top')
-ax2.set_xlabel('Epochs (Secondary X-Axis)')  # Label for the top x-axis
-ax2.yaxis.tick_right()
-ax2.yaxis.set_label_position("right")
-ax2.set_ylabel('Loss')  # Secondary y-axis label
-ax2.legend(loc='right')
-
-plt.show()
+#%% Fitting routine
+fig_path = utils.file_ID(pm.figs_dir,
+                          file_name(pm.architecture, net_ark, 'fitting'),
+                          pm.fig_format)
+# Target function
+target = lambda x: (1/np.pi)**(1/4) * np.exp(-0.5 * (x)**2)
+# Fitting
+utils.fitting(model, mesh, target, fig_path, visibility=True)
 
 #%% Stochastic Reconfiguration
 # Initial conditions
@@ -165,12 +107,12 @@ params = imag_evo.get_params()
 # Plot data
 fig_path = utils.file_ID(pm.figs_dir,
                          file_name(pm.architecture, net_ark, pm.evolution),
-                         ".png")
+                         pm.fig_format)
 plots.evo_fig_params(imag_evo.t_grid, mesh, den.T, params, fig_path=fig_path)
 
 #%% Stochastic Reconfiguration
 # Initial conditions
-pm.x0 = 0
+pm.x0 = 1
 pm.w = 1
 
 # Time parameters
@@ -198,10 +140,11 @@ params = real_evo.get_params()
 # Plot data
 fig_path = utils.file_ID(pm.figs_dir,
                          file_name(pm.architecture, net_ark, pm.evolution),
-                         ".png")
+                         pm.fig_format)
 plots.evo_fig_params(real_evo.t_grid, mesh, den.T, params, fig_path=fig_path)
 
 # %% Compare with analytic results
+
 # Analytic data
 x0 = 1
 p0 = 0
@@ -222,7 +165,7 @@ target_energy = 0.5 * (1 + p**2 + x**2)
 den_diff = den.T - target_den
 energy_diff = real_evo.energy - target_energy
 
-fig_path_0 = fig_path[:-4] + "_compare.png"
+fig_path_0 = fig_path + "_compare." + pm.fig_format
 plots.evo_fig_compare(real_evo.t_grid, mesh, den_diff, energy_diff, fig_path=fig_path_0)
 
 # %%
